backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
qdrant-client==1.11.1
open-clip-torch==2.26.1
torch==2.4.1
torchvision==0.19.1
Pillow==10.4.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
slowapi==0.1.9
